fix: Fix test order for and/or.
Split count to separate file.
